Proof of concept: Add speculation assertions on `provides`
`provide` and `require` avoid implicit dependencies between middlewares
which can be a problem in Rails controllers. We can take this further by
saying: not only does this middleware provide a particular value, but
its value conforms to a Speculation spec.

If Speculation's `check_asserts` configuration is enabled, any `provide`
call that has a relevant Speculation spec registered will have its value
checked for conformance. If the value doesn't conform to the spec, an
exception will be raised with details on why it doesn't conform.

Since Speculation uses a global spec registry with namespaced symbols,
one to tie together a `provides` declaration with a spec is to `provide`
the full namespaced spec name. This is a going to be a bit weird to
people not familiar with Speculation...